requirements.txt
py-cord==2.3.2
python-dotenv==0.21.1
beautifulsoup4==4.12.3
requests==2.32.3
tqdm==4.66.5